requirements_reel.txt
moviepy==1.0.3
Pillow==10.2.0
numpy==1.24.3
